add test/test.sh test/parse_main.c and relative file
